Issue #70 - Serialising Code lists to CSV-W (#97)
Code-lists are now serialised to CSV-W when processing a cube. This change includes a behave test which ensures the CSV-W is valid, that it converts to the appropriate RDF and that the RDF passes the SKOS SPARQL tests.

Also did a bit of refactoring in the QbWriter class to ensure that the URLs consistently point towards the CSV file instead of sometimes pointing to the metadata JSON file.