Got most (if not all) unit tests passing now.
Still need to add a couple of unit tests for text truncation (and anything else I've forgotten) and add a param to the CLI which allows the user to configure the depth of JSON schema validation errors which get written out.

Also still need to try out the behaviour tests and see what needs updating.