#583 Updating QB Models to specify observation value against attributes and units columns (#612)
This PR makes changes to the QbModels to enable linking the Attribute/Unit columns to their observation values. This means the QbAttribute and QbMultiUnits classes were changed so they can point back to the observed value column through the variable "observed_value_col_title".

Another significant change is that the classes QbSingleMeasureObservationValue and QbMultiMeasureObservationValue are now collapsed into one class named QbObservationValue, that used to simply contain abstract methods for its two subclasses, but now it is the concrete class that can differentiate between pivoted observation values or multi measure observation values through the optional "measure" component.

Any functions, tests, etc. that made use of the old QbSingleMeasureObservationValue and QbMultiMeasureObservationValue have been adjusted to instead check the measure of the QbObservationValue, to determine if it is dealing with a pivoted shape cube or a standard shape cube.

Co-authored-by: Gishan Don Ranasinghe <>
Co-authored-by: Nick Papoutsakis